1_face_mask_detection/requirements.txt
ultralytics>=8.0.0
PyYAML>=6.0
opencv-python>=4.8.0
